hierarchydemo: Implement experimental "copies" algorithm
Use a tree that is built in the beginning of the frame (considered needed for the gravity simulation to avoid duplicated work).
